[fix]チェックロジックを集約し、main-masterブランチに対応

+++ for_github/update_changelog.sh
@@ -9,7 +9,7 @@ tmpfile=tmp
 #--- 変更箇所はここまで
 
 # チェックロジック
-stop_flag=false
+stop_flag=false  # 処理を継続するか判断するためのフラグ。立ったら中断
 # 必要なコマンド（パッケージ）の確認
 if [ -z "$(which github_changelog_generator)" ]
 then
@@ -23,31 +23,6 @@ then
   stop_flag=true
 fi
 
-# リポジトリについては、引数で変更できるようにしておく
-if [ -z "$1" ]
-then
-  echo "[stop] Require repository"
-  stop_flag=true
-else
-  repository=$1
-fi
-
-if [ ${stop_flag} ]
-then
-  exit 1
-fi
-
-
-# .envがない、GITHUB_REPO_TOKENが設定されていない場合はユーザーに入力させる
-if [ -z "${GITHUB_REPO_TOKEN}" ]
-then
-  read -p "[input] github token(required permission repo):" GITHUB_REPO_TOKEN
-fi
-
-# 移動先: /タスクスケジューラ/changelog更新機/ファイル
-cd "$(dirname ${0})/../../"
-
-stop_flag=false  # 処理を継続するか判断するためのフラグ。立ったら中断
 if [ "$(git config user.name)" = "" ]
 then
   echo "[stop] Required `git config user.name.`"
@@ -66,12 +41,37 @@ then
   exit 1
 fi
 
-github_changelog_generator -u ${user} -p ${repository} -t ${GITHUB_REPO_TOKEN} --issues-label "### 終了・または先送りしたissue" --header-label "# 日付順" --unreleased-label "指定なし"
-github-changes -o ${user} -r ${repository} -k ${GITHUB_REPO_TOKEN} --use-commit-body -t "タグ別" -z Asia/Tokyo -m "YYYY年M月D日" -n "最終更新" -a -f tmp
-cat ${tmpfile} >>${file}
-rm ${tmpfile}
+# リポジトリについては、引数で変更できるようにしておく
+if [ -z "$1" ]
+then
+  echo "[stop] Require repository"
+  stop_flag=true
+else
+  repository=$1
+fi
+
+# .envがない、GITHUB_REPO_TOKENが設定されていない場合はユーザーに入力させる
+if [ -z "${GITHUB_REPO_TOKEN}" ]
+then
+  read -p "[input] github token(required permission repo):" GITHUB_REPO_TOKEN
+fi
+
+github_changelog_generator -u ${user} -p ${repository} -t ${GITHUB_REPO_TOKEN} --issues-label "### 終了・または先送りしたissue" --header-label "# 日付順" --unreleased-label "指定なし" -f ${file}
+github-changes -o ${user} -r ${repository} -k ${GITHUB_REPO_TOKEN} --use-commit-body -main -t "タグ別" -z Asia/Tokyo -m "YYYY年M月D日" -n "最終更新" -a -f ${tmpfile} 2>/dev/null
+
+# 上記github-changesで失敗した場合は古いブランチ(master)の可能性があるので再試行する
+if [ ! -f "${file}" ]
+then
+  github-changes -o ${user} -r ${repository} -k ${GITHUB_REPO_TOKEN} -b master --use-commit-body -t "タグ別" -z Asia/Tokyo -m "YYYY年M月D日" -n "最終更新" -a -f ${tmpfile}
+fi
+
+if [ -f "${file}" ]
+then
+  cat ${tmpfile} >>${file}
+  rm ${tmpfile}
+fi
 
 git add ${file}
 git commit -m "update: ${file}"
-git tag "v${date '+v%Y/%m/%d'}"
+git tag "${date '+v%Y/%m/%d'}"
 git push --tags
